abudeeva/requirements.txt
python-slugify
django-mptt
django-mptt-admin
pillow
easy-thumbnails